+++ app/functions-extras.php
@@ -81,3 +81,23 @@ add_filter( 'backdrop/view/content/data', function( $data ) {
 	return $data;
 
 } );
+
+/**
+ * Filters the excerpt more link.
+ *
+ * @since  1.0.0
+ * @access public
+ * @return string
+ */
+add_filter( 'excerpt_more', function() {
+
+	return sprintf(
+		'&thinsp;&hellip;&thinsp;<a href="%s" class="entry-more-link">%s</a>',
+		esc_url( get_permalink() ),
+		sprintf(
+			// Translators: %s is the post title for screen readers.
+			esc_html__( 'Continue reading&nbsp;%s&nbsp;&rarr;', 'amicable' ),
+			the_title( '<span class="screen-reader-text">', '</span>', false )
+		)
+	);
+} );